t/base64-encoding/codificacion-base64 closes #16
