+++ public/huellas-digitales-tipo-sha1
@@ -193,7 +193,7 @@ debes investigar cuidadosamente sobre
       
       
       <p class="next">
-        Siguiente ejemplo: <a href="base64-encoding">Base64 Encoding</a>.
+        Siguiente ejemplo: <a href="codificacion-base64">Codificación Base64</a>.
       </p>
       
       <p class="footer">
