+++ public/index.html
@@ -143,7 +143,7 @@
       
         <li><a href="huellas-digitales-tipo-sha1">Huellas digitales tipo SHA1</a></li>
       
-        <li><a href="base64-encoding">Base64 Encoding</a></li>
+        <li><a href="codificacion-base64">Codificación Base64</a></li>
       
         <li><a href="reading-files">Reading Files</a></li>
